requirements.txt
bcrypt==3.2.2
channels==3.0.5
DateTime==4.5
Django==3.0
django-cors-headers==3.13.0
mysql==0.0.3
mysqlclient==2.1.1
PyJWT==2.4.0
requests==2.28.1
sqlparse==0.4.2
